dds2val/requirements/requirements.txt
# feeder
cyclonedds
kuksa-client
# mapper
pyyaml
py-expression-eval
